requirements.txt
ollama==0.1.6
pandas==2.2.0
pillow==10.2.0
